fix splitter build. correct libraries usage for blobs too.
