+++ ydb/core/tx/columnshard/common/ya.make
@@ -7,6 +7,7 @@ SRCS(
     snapshot.cpp
     portion.cpp
     tablet_id.cpp
+    blob.cpp
 )
 
 PEERDIR(
